requirements.txt
requests~=2.28.2
pytest
voluptuous~=0.13.1
pytest-voluptuous
selene==2.0.0b16
allure-pytest~=2.12.0
python-dotenv
